pyproject.toml
[tool.poetry]
name = "py-fastapi-online-cinema"
version = "0.1.0"
description = "Online Cinema API with FastAPI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.110.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
uvicorn = ">=0.27.0"
sqlalchemy = "^2.0.0"
alembic = "^1.11.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
celery = "^5.3.0"
redis = "^5.0.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
